requirements.txt
fastapi==0.109.0
sqlmodel==0.0.14
uvicorn[standard]==0.27.0
requests==2.31.0
python-multipart==0.0.6
python-dotenv==1.0.0